chore: move app files back to root
