+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import App from "./components/App";
+import App from "./App";
 import { GlobalStyle } from "./styles/global";
 
 createRoot(document.getElementById("root")!).render(
